feat: move guide config
