+++ apps/builder/src/components/Guide/index.tsx
@@ -5,8 +5,8 @@ import { FC, HTMLAttributes, useMemo } from "react"
 import { useDispatch, useSelector } from "react-redux"
 import { Button } from "@illa-design/react"
 import { GuidePopover } from "@/components/Guide/GuidePopover"
-import { GUIDE_STEP } from "@/components/Guide/config"
 import { applyGuideStyle, stepMaskStyle } from "@/components/Guide/style"
+import { GUIDE_STEP } from "@/config/guide/config"
 import { getCachedAction } from "@/redux/config/configSelector"
 import { getCurrentStep } from "@/redux/guide/guideSelector"
 import { guideActions } from "@/redux/guide/guideSlice"
