+++ apps/builder/src/redux/guide/guideSelector.ts
@@ -1,5 +1,5 @@
 import { createSelector } from "@reduxjs/toolkit"
-import { GUIDE_STEP } from "@/components/Guide/config"
+import { GUIDE_STEP } from "@/config/guide/config"
 import { getCurrentId, getTeamItems } from "@/redux/team/teamSelector"
 import { RootState } from "@/store"
